lengkapin

+++ application/controllers/menu.php
@@ -7,6 +7,7 @@ class menu extends CI_controller
         parent::__construct();
         is_logged_in();
         $this->load->library('form_validation');
+        $this->load->model('m_master_data');
     }
     public function index()
     {
@@ -24,8 +25,8 @@ class menu extends CI_controller
             $this->main_generic->layout($pages, $data);
         } else {
             $this->db->insert('user_menu', ['menu' => $_POST['menu']]);
-            $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">
-			Congratulation! your data has been saved </div>');
+            $alert = $this->main_generic->alert('Berhasil', 'Data berhasil ditambahkan', 'success');
+            $this->session->set_flashdata('message', $alert);
             redirect('menu');
         }
     }
@@ -58,20 +59,48 @@ class menu extends CI_controller
             redirect('menu/submenu');
         }
     }
+    public function edit_menu($id){
+        $pages  = 'menu/edit_menu';
+        $data['user']           = $this->db->get_where('user', ['email' => $this->session->userdata('email')])->row_array();
+        $data['menu']           = $this->db->get_where('user_menu', ['id' => $id])->row();
+        $data['title']          = 'Edit menu';
+        $data['subsubtitle']    = 'Fill the form';
+        $this->form_validation->set_rules('menu', 'Menu', 'required', [
+            'required' => 'Kolom %s harus diisi'
+        ]);
+        if ($this->form_validation->run() == false) {
+            $this->main_generic->layout($pages, $data);
+        } else {
+            $this->db->set('menu', $_POST['menu']);
+            $this->db->where('id', $id);
+            $this->db->update('user_menu');
+            $alert = $this->main_generic->alert('Berhasil', 'Data berhasil ditambahkan', 'success');
+            $this->session->set_flashdata('message', $alert);
+            redirect('menu');
+        }
+    }
     public function deleteData($id)
     {
-        $this->db->where('id', $id);
-        $this->db->delete('user_menu');
-        $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">
-        Congratulation! your data has been Deleted </div>');
+        $cari_sub_menu = $this->m_master_data->cari_sub_menu_by_user_menu($id);
+        if($cari_sub_menu == null){
+            $this->db->where('id', $id);
+            $this->db->delete('user_menu');
+            $alert = $this->main_generic->alert('Berhasil', 'Data berhasil didelete', 'success');
+            $this->session->set_flashdata('message', $alert);
+        } else {
+            $alert = $this->main_generic->alert('Gagal', 'Data masih tersedia di sub menu', 'danger');
+            $this->session->set_flashdata('message', $alert);
+        }
         redirect('menu');
     }
     public function deleteDataSubmenu($id)
     {
         $this->db->where('id', $id);
         $this->db->delete('user_sub_menu');
-        $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">
-        Congratulation! your data has been Deleted </div>');
+        $alert = $this->main_generic->alert('Berhasil', 'Data berhasil didelete', 'success');
+        $this->session->set_flashdata('message', $alert);
+        // $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">
+        // Congratulation! your data has been Deleted </div>');
         redirect('menu/submenu');
     }
 }
